requirements.txt

langchain==0.3.10
python-dotenv>=0.20.0
langgraph>=0.1.0
beeai-framework>=0.1.14
langflow==1.1.1
langchain_community==0.3.10
langchain-ibm
flask
